Downgrade HikariCP version to 3.4.3
HikariCP 3.4.5 is a multi-release jar, which contains a module descriptor for Java 11 (and up). Unfortunately this module descriptor refers to the module slf4j.api, which is the filename-derived automatic module name of older versions of slf4j-api. Spring Boot 2.3 however brings in a newer version of slf4j, which declares the automatic module name org.slf4j in its manifest. This leads to failure to load this module, which in turn leads to Spring autoconfiguration for Hikari not loading, causing a chain reaction which leads to the EntityManager not being loaded. Therefore revert to the latest version of HakiriCP that does not contain the offending module descriptor (3.4.3).